style connect modal (consistent with staking)
