+++ src/ui/components/accounts/Account.tsx
@@ -1,5 +1,6 @@
 import Identicon from '@polkadot/react-identicon';
 import { stringShorten } from '@polkadot/util';
+import { ConnectCard } from './ConnectCard';
 
 export interface AccountProps {
   name: string;
@@ -11,19 +12,19 @@ export interface AccountProps {
 const Account = ({ name, address, clickHandler, state }: AccountProps) => {
   const { isConnected } = state;
   return (
-    <div
-      className="mx-1 flex flex-nowrap rounded-lg border border-solid"
+    <ConnectCard
+      className="flex flex-nowrap p-1"
       onClick={() => clickHandler()}
     >
       <div className="m-1 flex items-center justify-center">
-        <Identicon value={address} theme="polkadot" size={32} />
+        <Identicon value={address} theme="polkadot" size={24} />
       </div>
-      <div className="m-1 flex flex-auto flex-col">
+      <div className="m-1 flex flex-auto flex-col text-xs">
         <div>{stringShorten(name, 10)}</div>
         <div>{stringShorten(address, 10)}</div>
       </div>
       <div className="m-1">{isConnected ? 'connected' : ''}</div>
-    </div>
+    </ConnectCard>
   );
 };
