+++ src/ui/components/accounts/ConnectCard.tsx
@@ -1,0 +1,18 @@
+import { ClickableProps } from 'src/ui/lib/types';
+
+export function ConnectCard({
+  className,
+  children,
+  style,
+  onClick,
+}: ClickableProps): JSX.Element {
+  return (
+    <div
+      className={`rounded-lg bg-gray-100 ${className || ''}`}
+      style={style}
+      onClick={onClick}
+    >
+      {children}
+    </div>
+  );
+}
