+++ src/ui/components/index.tsx
@@ -1,4 +1,4 @@
-export { ConnectButton } from './Connect.js';
+export { ConnectButton } from './accounts/ConnectButton.js';
 export { ErrorBoundary } from './ErrorBoundary.js';
 export { Header } from './Header.js';
 export { LoadingPanel } from './Loading.js';
